# docs/openapi.yaml
openapi: 3.1.0
info:
  title: GenXcare API
  version: 1.0.0
  description: REST contract for the GenXcare modular monolith.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/seniors/{seniorId}/appointments:
    parameters: [{ $ref: "#/components/parameters/SeniorId" }]
    get:
      tags: [Appointments]
      operationId: listAppointments
      parameters: [{ $ref: "#/components/parameters/Scope" }, { $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { $ref: "#/components/responses/Page" }, default: { $ref: "#/components/responses/Error" } }
    post:
      tags: [Appointments]
      operationId: createAppointment
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses: { "201": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
  /v1/appointments/{appointmentId}:
    parameters: [{ $ref: "#/components/parameters/AppointmentId" }]
    get:
      tags: [Appointments]
      operationId: getAppointment
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
    patch:
      tags: [Appointments]
      operationId: updateAppointment
      requestBody: { $ref: "#/components/requestBodies/JsonObject" }
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
  /v1/appointments/{appointmentId}/cancel:
    parameters: [{ $ref: "#/components/parameters/AppointmentId" }]
    post:
      tags: [Appointments]
      operationId: cancelAppointment
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
  /v1/appointments/{appointmentId}/complete:
    parameters: [{ $ref: "#/components/parameters/AppointmentId" }]
    post:
      tags: [Appointments]
      operationId: completeAppointment
      responses: { "200": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
...
  /v1/seniors/{seniorId}/messages:
    parameters: [{ $ref: "#/components/parameters/SeniorId" }]
    get:
      tags: [Messages]
      operationId: listMessages
      parameters: [{ $ref: "#/components/parameters/Cursor" }, { $ref: "#/components/parameters/Limit" }]
      responses: { "200": { $ref: "#/components/responses/MessagePage" }, default: { $ref: "#/components/responses/Error" } }
    post:
      tags: [Messages]
      operationId: createMessage
      requestBody: { $ref: "#/components/requestBodies/MessageContent" }
      responses: { "201": { $ref: "#/components/responses/Entity" }, default: { $ref: "#/components/responses/Error" } }
  /v1/seniors/{seniorId}/messages/read:
    parameters: [{ $ref: "#/components/parameters/SeniorId" }]
    post:
      tags: [Messages]
      operationId: markMessagesRead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [throughMessageId]
              properties: { throughMessageId: { $ref: "#/components/schemas/Uuid" } }
      responses: { "204": { description: Read state advanced }, default: { $ref: "#/components/responses/Error" } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    SeniorId: { name: seniorId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
...
    AppointmentId: { name: appointmentId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
    Cursor: { name: cursor, in: query, schema: { type: string } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
    Scope: { name: scope, in: query, schema: { type: string } }
  requestBodies:
    JsonObject:
      required: true
      content: { application/json: { schema: { type: object } } }
...
    MessageContent:
      required: true
      content:
        application/json:
          schema:
            type: object
            additionalProperties: false
            required: [content]
            properties: { content: { type: string, minLength: 1, maxLength: 2000 } }
  responses:
...
    Entity:
      description: Resource
      content: { application/json: { schema: { type: object } } }
...
    Page:
      description: Cursor page
      content: { application/json: { schema: { $ref: "#/components/schemas/PageEnvelope" } } }
    MessagePage:
      description: Cursor page of messages with unread state
      content: { application/json: { schema: { $ref: "#/components/schemas/MessagePage" } } }
    Error:
      description: Request failed
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
    Uuid: { type: string, format: uuid }
    ListEnvelope:
      type: object
      required: [items]
      properties: { items: { type: array, items: { type: object } } }
    PageEnvelope:
      allOf:
        - { $ref: "#/components/schemas/ListEnvelope" }
        - { type: object, required: [nextCursor], properties: { nextCursor: { type: [string, "null"] } } }
    MessagePage:
      allOf:
        - { $ref: "#/components/schemas/PageEnvelope" }
        - { type: object, required: [unreadCount], properties: { unreadCount: { type: integer, minimum: 0 } } }
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: { type: object, additionalProperties: { type: string } }